% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lookupTable.R
\docType{methods}
\name{initialize,lookupTable-method}
\alias{initialize,lookupTable-method}
\title{Initialize and construct a lookupTable object}
\usage{
\S4method{initialize}{lookupTable}(.Object, df.input, response,
  feature.boundaries, features.con = character(0),
  features.cat = character(0), fill.method = "mean")
}
\arguments{
\item{.Object}{the prototype object}

\item{df.input}{training data set containing columns with names found in features.con and features.cat vectors}

\item{response}{name of the response variable}

\item{feature.boundaries}{a list of thresholds for each continuous feature (names contained in feature.con) to construct bins. Should use -Inf and Inf as the first and last values, respectively.}

\item{features.con}{a vector of continuous feature names}

\item{features.cat}{a vector of categorical feature names}

\item{fill.method}{the method to fill entries of the table ('mean' or 'median')}
}
\value{
A lookupTable object with a table trained with df.input data
}
\description{
Initialize and construct a lookupTable object
}

