% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lookupTable.R
\docType{class}
\name{lookupTable-class}
\alias{lookupTable}
\alias{lookupTable-class}
\title{An S4 class that defines the look-up table and all other components required for prediction using this table.}
\description{
An S4 class that defines the look-up table and all other components required for prediction using this table.
}
\section{Slots}{

\describe{
\item{\code{table}}{the look-up table with entries to be retrieved as prediction results}

\item{\code{feature.con}}{a vector of continuous feature names}

\item{\code{feature.cat}}{a vector of categorical feature names}

\item{\code{feature.boundaries}}{a list of boundaries for each input feature (inferred during construction from input data)}

\item{\code{response}}{the name of the response variable for the look-up table}

\item{\code{default}}{the default value for cells corresponding to a missing combination of input values}

\item{\code{response.categories}}{sequence of all categories (order-dependent) for the response variable, if it's categorical}
}}

